+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 0.7.10
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Implement annotations @PostConstruct and @PreDestroy for session and message beans
+* Implement @Startup annotation for singleton session beans
+
 # Version 0.7.9
 
 ## Bugfixes